+++ bc/core/src/main/java/app/giftify/order/adapter/outbound/batch/ExpirationJobConfig.java
@@ -1,8 +1,8 @@
 package app.giftify.order.adapter.outbound.batch;
 
 import lombok.RequiredArgsConstructor;
-import org.springframework.batch.core.Job;
-import org.springframework.batch.core.Step;
+import org.springframework.batch.core.job.Job;
+import org.springframework.batch.core.step.Step;
 import org.springframework.batch.core.job.builder.JobBuilder;
 import org.springframework.batch.core.repository.JobRepository;
 import org.springframework.context.annotation.Bean;
