+++ bc/core/src/main/java/app/giftify/order/adapter/outbound/batch/ExpirationJobListener.java
@@ -1,9 +1,9 @@
 package app.giftify.order.adapter.outbound.batch;
 
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.batch.core.JobExecution;
-import org.springframework.batch.core.JobExecutionListener;
-import org.springframework.batch.core.StepExecution;
+import org.springframework.batch.core.job.JobExecution;
+import org.springframework.batch.core.listener.JobExecutionListener;
+import org.springframework.batch.core.step.StepExecution;
 import org.springframework.stereotype.Component;
 
 import java.time.Duration;
