+++ bc/settlement/src/main/java/app/giftify/settlement/adapter/outbound/batch/execution/ExecutionItemConfig.java
@@ -3,7 +3,7 @@ package app.giftify.settlement.adapter.outbound.batch.execution;
 import app.giftify.settlement.application.service.SettlementExecutionService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.batch.core.configuration.annotation.StepScope;
-import org.springframework.batch.item.support.ListItemReader;
+import org.springframework.batch.infrastructure.item.support.ListItemReader;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
