+++ bc/settlement/src/main/java/app/giftify/settlement/adapter/outbound/batch/execution/SellerIdPartitioner.java
@@ -5,8 +5,8 @@ import app.giftify.settlement.application.outbound.port.SettlementQueueRepositor
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.batch.core.configuration.annotation.StepScope;
-import org.springframework.batch.core.partition.support.Partitioner;
-import org.springframework.batch.item.ExecutionContext;
+import org.springframework.batch.core.partition.Partitioner;
+import org.springframework.batch.infrastructure.item.ExecutionContext;
 import org.springframework.stereotype.Component;
 import org.springframework.util.CollectionUtils;
