+++ bc/settlement/src/main/java/app/giftify/settlement/adapter/outbound/batch/execution/SettlementExecutionProcessor.java
@@ -2,7 +2,7 @@ package app.giftify.settlement.adapter.outbound.batch.execution;
 
 import app.giftify.settlement.application.service.SettlementExecutionService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.batch.item.ItemProcessor;
+import org.springframework.batch.infrastructure.item.ItemProcessor;
 
 @RequiredArgsConstructor
 public class SettlementExecutionProcessor implements ItemProcessor<Long, ExecutionResult> {
