+++ bc/settlement/src/main/java/app/giftify/settlement/adapter/outbound/batch/execution/SettlementExecutionWriter.java
@@ -5,8 +5,8 @@ import app.giftify.shared.api.exception.InfraErrorCode;
 import app.giftify.shared.api.exception.InfraException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.batch.item.Chunk;
-import org.springframework.batch.item.ItemWriter;
+import org.springframework.batch.infrastructure.item.Chunk;
+import org.springframework.batch.infrastructure.item.ItemWriter;
 
 @Slf4j
 @RequiredArgsConstructor
