+++ bc/settlement/src/main/java/app/giftify/settlement/adapter/outbound/batch/validation/OrderIdPartitioner.java
@@ -5,8 +5,8 @@ import app.giftify.settlement.application.outbound.port.SettlementItemRepository
 import app.giftify.settlement.domain.status.SettlementItemStatus;
 import lombok.RequiredArgsConstructor;
 import org.springframework.batch.core.configuration.annotation.StepScope;
-import org.springframework.batch.core.partition.support.Partitioner;
-import org.springframework.batch.item.ExecutionContext;
+import org.springframework.batch.core.partition.Partitioner;
+import org.springframework.batch.infrastructure.item.ExecutionContext;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
