+++ bc/settlement/src/main/java/app/giftify/settlement/adapter/outbound/batch/validation/ValidationBatchScheduler.java
@@ -3,9 +3,9 @@ package app.giftify.settlement.adapter.outbound.batch.validation;
 import app.giftify.settlement.adapter.outbound.batch.common.BatchProperties;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
-import org.springframework.batch.core.Job;
-import org.springframework.batch.core.JobParameters;
-import org.springframework.batch.core.JobParametersBuilder;
+import org.springframework.batch.core.job.Job;
+import org.springframework.batch.core.job.parameters.JobParameters;
+import org.springframework.batch.core.job.parameters.JobParametersBuilder;
 import org.springframework.batch.core.launch.JobLauncher;
 import org.springframework.scheduling.annotation.EnableScheduling;
 import org.springframework.scheduling.annotation.Scheduled;
