+++ bc/settlement/src/main/java/app/giftify/settlement/adapter/outbound/batch/validation/ValidationJobListener.java
@@ -2,8 +2,8 @@ package app.giftify.settlement.adapter.outbound.batch.validation;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.batch.core.BatchStatus;
-import org.springframework.batch.core.JobExecution;
-import org.springframework.batch.core.JobExecutionListener;
+import org.springframework.batch.core.job.JobExecution;
+import org.springframework.batch.core.listener.JobExecutionListener;
 import org.springframework.stereotype.Component;
 
 @Component
